refactor: split and relocate MsgPack helpers into a dedicated helpers file
